Feat/schedules

+++ src/schedules/schedules.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, BadRequestException } from '@nestjs/common';
 import { CreateScheduleDto } from './dto/create-schedule.dto';
 import { UpdateScheduleDto } from './dto/update-schedule.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -255,6 +255,17 @@ export class SchedulesService {
           await this.ScheduleClassRepository.save(scheduleClasses);
         }
       }
+
+      responseData.schedules = responseData.schedules.map((schedule: any) => {
+        schedule.classes = schedule.classes.map((cls: any) => {
+          const matchedCourse = courses.find((c) => c.course_id === cls.course_id);
+          return {
+            ...cls,
+            course_name: matchedCourse ? matchedCourse.course_name : null,
+          };
+        });
+        return schedule;
+      });
     }
 
     return responseData;
@@ -360,6 +371,14 @@ export class SchedulesService {
 
   async remove(student_id: string, id: number) {
     const schedule = await this.findOne(student_id ,id);
+
+    if (schedule.is_selected) {
+      throw new BadRequestException({
+        success: false,
+        error: { code: 'SCHEDULE_IS_SELECTED', message: `Không thể xóa thời khóa biểu đã được chọn` }
+      });
+    }
+
     return await this.ScheduleRepository.remove(schedule);
   }
 }
